+++ apps/bbs/pages/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Head from 'next/head';
 import PropTypes from 'prop-types';
 import 'antd/dist/antd.css';
-import DefaultCom from './layout/default/default.js';
+import DefaultCom from './layout/default/default';
 
 class Home extends Component {
     render() {
