--- apps/bbs/pages/layout/default/default.js
@@ -1,39 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Layout, Menu, Breadcrumb } from 'antd';
-const { Header, Content, Footer } = Layout;
-
-export default class DefaultCom extends React.Component {
-    render() {
-        console.log('this.props', this.props);
-        console.log('this.context', this.context);
-        return (
-            <Layout className="layout">
-                <Header style={{backgroundColor: '#f44336', boxShadow: '0 2px 5px 0 rgba(0,0,0,.26)'}}>
-                    <div className="logo" />
-                    <Menu
-                        theme="dark"
-                        mode="horizontal"
-                        defaultSelectedKeys={['2']}
-                        style={{ lineHeight: '64px' }}
-                    >
-                        <Menu.Item key="1">nav 1</Menu.Item>
-                        <Menu.Item key="2">nav 2</Menu.Item>
-                        <Menu.Item key="3">nav 3</Menu.Item>
-                    </Menu>
-                </Header>
-                <Content style={{ padding: '0 50px' }}>
-                <Breadcrumb style={{ margin: '16px 0' }}>
-                    <Breadcrumb.Item>Home</Breadcrumb.Item>
-                    <Breadcrumb.Item>List</Breadcrumb.Item>
-                    <Breadcrumb.Item>App</Breadcrumb.Item>
-                </Breadcrumb>
-                <div style={{ background: '#fff', padding: 24, minHeight: 280 }}>Content</div>
-                </Content>
-                <Footer style={{ textAlign: 'center' }}>
-                Ant Design ©2018 Created by Ant UED
-                </Footer>
-            </Layout>
-        )
-    }
-}
