+++ apps/bbs/pages/layout/default/default.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Layout, Menu, Breadcrumb } from 'antd';
+import FooterCom from './footer';
+import HeaderCom from './head';
+const { Header, Content } = Layout;
+
+export default class DefaultCom extends React.Component {
+    render() {
+        return (
+            <Layout className="layout">
+                <HeaderCom />
+                <Content style={{ padding: '0 50px' }}>
+                    <Breadcrumb style={{ margin: '16px 0' }}>
+                        <Breadcrumb.Item>Home</Breadcrumb.Item>
+                        <Breadcrumb.Item>List</Breadcrumb.Item>
+                        <Breadcrumb.Item>App</Breadcrumb.Item>
+                    </Breadcrumb>
+                    <div style={{ background: '#fff', padding: 24, minHeight: 280 }}>Content</div>
+                </Content>
+                <FooterCom />
+            </Layout>
+        )
+    }
+}
